Mark the difference between Parameter and argument
In Jonathan's perl 6 introduction, he specifically wanted to stress the
difference between Parameter (a thing you define in a sub/method/pointy),
a Signature (zero or more parameters of a sub/method/pointy), an
argument (something you pass to a sub/method/pointy) and an argument
list or Capture.  And that error messages try to adhere to this idiom.

So far, in this perl 6 intro, only arguments were mentioned.