Finish Encapsulation
For the record: I've literally translated the english text, but feel
that the english text is taking the wrong approach to explaining these
concepts.  For one, it seems to assume that you can initialize private
attributes with the default .new().  This is incorrect.  Having a '.'
twigil on an attribute really means 2 things:

 1. generate an accessor if there is no method defined for it
 2. allow initialization using the default .new (more technically,
    the default .bless method).